+++ src/main/java/com/zhongkexinli/micro/serv/common/util/ForEachUtils.java
@@ -1,8 +1,11 @@
 package com.zhongkexinli.micro.serv.common.util;
 
 
+import java.util.List;
 import java.util.Objects;
 import java.util.function.BiConsumer;
+import java.util.function.Consumer;
+import java.util.function.IntFunction;
 
 /**
  * 
@@ -34,6 +37,22 @@ public class ForEachUtils {
             }
             
             action.accept(index-1, element);
+        }
+    }
+    
+    public static void foreach(List<?> list, int batchSize, Consumer<List<?>> consumer) {
+        for (int i = 0; i < list.size(); i += batchSize) {
+            consumer.accept(list.subList(i, Math.min(list.size(), i + batchSize)));
+        }
+    }
+
+    public static void foreach(IntFunction<Integer> function) {
+        foreach(0, function);
+    }
+
+    public static void foreach(int start, IntFunction<Integer> function) {
+        for (int end = start + 1; start < end; start++) {
+            end = function.apply(start);
         }
     }
 }