+++ src/test/java/com/zhongkexinli/micro/serv/common/util/ForEachUtilsTest.java
@@ -1,0 +1,17 @@
+package com.zhongkexinli.micro.serv.common.util;
+
+import org.junit.Test;
+
+public class ForEachUtilsTest {
+
+    @Test
+    public void foreachTest() {
+        
+        int forSize = 10;
+        ForEachUtils.foreach( a -> {
+           System.out.println("do aaa:"+a);
+           return forSize;
+        });
+    }
+
+}